% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_select.R
\name{delete_select}
\alias{delete_select}
\title{Selecionando series para delete}
\usage{
delete_select(
  indicator = as.character(),
  new_sids = as.character(),
  token = as.character(),
  url = as.character()
)
}
\description{
Selecionando series para delete
}
\examples{
\dontrun{
}

}
\author{
Luiz Paulo Tavares Gonçalves
}
