% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clean_sids.R
\name{get_clean_sids}
\alias{get_clean_sids}
\title{Importando os Sids de todas as series do indicador selecionado}
\usage{
get_clean_sids(metadados_series = as.list())
}
\description{
Importando os Sids de todas as series do indicador selecionado
}
\examples{
\dontrun{

}

}
\author{
Luiz Paulo Tavares Gonçalves
}
