% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.r
\name{get_id}
\alias{get_id}
\title{ID dos nós da árvore}
\usage{
get_id(tree = as.list(), node = as.character())
}
\arguments{
\item{\code{tree}}{representa uma lista com a árvore da FS}

\item{\code{node}}{representa o nó com seus respectivos filhos que deseja acessar}
}
\value{
O retorno é Tibble os IDs do nós e filhos selecionados
}
\description{
A função retorna os IDs dos nós e dos filhos da árvore da FS
}
\examples{

\dontrun{

id_tree_fs = FScuradoria::get_id(tree = tree_fs, node = "Brazil")

}

}
\author{
Gabriel Bellé
}
