% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadados_series.R
\name{get_metadados_series}
\alias{get_metadados_series}
\title{Importando os metadados de todas as series do indicador selecionado}
\usage{
get_metadados_series(
  indicator = as.character(),
  token = as.character(),
  url = as.character()
)
}
\description{
Importando os metadados de todas as series do indicador selecionado
}
\examples{
\dontrun{

metadados_series = get_metadados_series(indicator = indicator,
                                        token = token_to_use,
                                        url = url_to_use)

}

}
\author{
Luiz Paulo Tavares Gonçalves
}
